Implement first michel cut.
This cut also, crucially, begins to construct the MichelEvent object.

I would like to change some naming here. And another todo is to NOT
NOT modify the event object, but to return it instead a la functional
programming.

Currently not working in my setup because I don't have the needed CVU
functions.

TODO: make such functions in a LowRecoilPionFunctions file. Add the
one other cut.