Save the michel cluster index to the michel object.
We need the cluster index in order to fix the double counting between
tracked and untracked clusters.

Other changes: switch a few hardcoded constants to use PlotUtilsPhysicalConstants.
Reorder member variables to better resemble the order that they're initialized.
(which is best practice in C++ and can lead to bugs otherwise).